--- Assets/01.Scenes/HUD-test.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 217b60a6bb9dd29418f6333877202162
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
