--- Assets/01.Scenes/UI-HUD-Test.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 59fd4b89006016b4fa949d3018073657
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
