--- Assets/02.Scripts/Inventory-Item/SlotDelete.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class SlotDelete : ButtonClicked
-{
-    public InventoryManager inventoryManager;
-    public GameObject Slot;
-    public override void ButtonEvent(PointerEventData eventData)
-    {
-        inventoryManager.removeItem(Slot.transform.GetSiblingIndex());
-    }
-}
