+++ Assets/02.Scripts/Inventory-Item/bcSlotDelete.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c63c885767dc87b4da75855f67af1830
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
