+++ Assets/02.Scripts/Inventory-Item/bcSlotSelect.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.EventSystems;
+
+public class bcSlotSelect : ButtonClicked
+{
+    public InventoryManager inventoryManager;
+
+    public override void ButtonEvent(PointerEventData eventData)
+    {
+        if(eventData.button == PointerEventData.InputButton.Left)
+        {
+            inventoryManager.SetPrimary(transform.GetSiblingIndex());
+        }
+        if(eventData.button == PointerEventData.InputButton.Right)
+        {
+            inventoryManager.SetSecondary(transform.GetSiblingIndex());
+        }
+    }
+}
