+++ Assets/02.Scripts/Inventory-Item/bcSlotSelect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ef53a228a777044f95be659fb3e7547
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
