--- Assets/02.Scripts/UI-MainMenu/ApplySettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eaf7611a9036a8649bc26e8af6f9ea7d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
