--- Assets/02.Scripts/UI-MainMenu/MenuLeft.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.Events;
-
-public class MenuLeft : ButtonClicked
-{
-    public MapSelect mapSelect;
-    public override void ButtonEvent(PointerEventData eventData)
-    {
-         mapSelect.index -= 1;
-    }
-}
