--- Assets/02.Scripts/UI-MainMenu/MenuLeft.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3e5d2ac8058b63f40aedb260b94f24de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
