--- Assets/02.Scripts/UI-MainMenu/MenuRight.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: af1faf85b0ebfae4497a614a59ebb819
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
