--- Assets/02.Scripts/UI-MainMenu/MyMenuButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 04d81b9c03c567643acc41b7edcee15b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
