--- Assets/02.Scripts/UI-MainMenu/MyMenuButtonGroup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7fcc5b48a1d2d8142bfeb2f7c06b1508
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
