--- Assets/02.Scripts/UI-MainMenu/OptionClicked.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b5a760c2d293c144bbfb470d5342969e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
