+++ Assets/02.Scripts/UI-MainMenu/bcMainMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4672ff2d547f18438ed0684a3ca057f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
