+++ Assets/02.Scripts/UI-MainMenu/bcMenuLeft.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.EventSystems;
+using UnityEngine.Events;
+
+public class bcMenuLeft : ButtonClicked
+{
+    public MapSelect mapSelect;
+    public override void ButtonEvent(PointerEventData eventData)
+    {
+         mapSelect.index -= 1;
+    }
+}
