+++ Assets/03.Prefabs/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2442f81db5322f7468063394f16a431d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
