+++ Assets/06.Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2304a515a36e6ba44974ad54ee517a58
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
